refactor(Utils): abstract Expirable from Cache
