+++ src/utils/index.js
@@ -1,5 +1,4 @@
 import isPlainObject from 'lodash/isPlainObject';
-import { default as d } from 'debug';
 
 export const tryAll = promiseConstructors => {
   const promise = new Promise(promiseConstructors[0]);
@@ -11,38 +10,8 @@ export const tryAll = promiseConstructors => {
 
 export const tap = interceptor => value => ((interceptor(value), value));
 
-const debug = d('LC:Cache');
-export class Cache {
-  constructor(name = 'anonymous') {
-    this.name = name;
-    this._map = {};
-  }
-
-  get(key) {
-    const cache = this._map[key];
-    if (cache) {
-      const expired = cache.expiredAt && cache.expiredAt < Date.now();
-      if (!expired) {
-        debug(`[${this.name}] hit: ${key} ${cache.value}`);
-        return cache.value;
-      }
-      debug(`[${this.name}] expired: ${key}`);
-      delete this._map[key];
-    }
-    debug(`[${this.name}] missed: ${key}`);
-    return null;
-  }
-
-  set(key, value, ttl) {
-    debug(`[${this.name}] set: ${key} ${value} ${ttl}`);
-    const cache = this._map[key] = {
-      value,
-    };
-    if (typeof ttl === 'number') {
-      cache.expiredAt = Date.now() + ttl;
-    }
-  }
-}
+export { default as Expirable } from './expirable';
+export { default as Cache } from './cache';
 
 export const decodeDate = date => {
   if (!date) return date;
